rename: 패키지 구조 변경
DTO 패키지 내 구조 가독성을 위해 DTO 패키지 안에 또 별도로 도메인별로 나눴습니다
