+++ src/main/java/com/yongfill/server/domain/posts/dto/like/LikeDto.java
@@ -1,0 +1,4 @@
+package com.yongfill.server.domain.posts.dto.like;
+
+public class LikeDto {
+}
