+++ src/main/java/com/yongfill/server/domain/posts/dto/post/DeletePostDto.java
@@ -1,4 +1,4 @@
-package com.yongfill.server.domain.posts.dto;
+package com.yongfill.server.domain.posts.dto.post;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -10,21 +10,7 @@ import org.springframework.http.HttpStatus;
 import java.util.List;
 
 @Builder
-public class CreatePostDto {
-
-    @Data
-    @Builder
-    @AllArgsConstructor
-    public static class RequestDto {
-        @NotBlank
-        private String title;
-        @NotBlank
-        private String category;
-        @NotBlank
-        private String content;
-
-    }
-
+public class DeletePostDto {
 
     @Builder
     @Getter
