+++ src/main/java/com/yongfill/server/domain/posts/service/PostServiceImpl.java
@@ -2,10 +2,10 @@ package com.yongfill.server.domain.posts.service;
 
 import com.yongfill.server.domain.member.entity.Member;
 import com.yongfill.server.domain.member.repository.MemberJpaRepository;
-import com.yongfill.server.domain.posts.dto.CreatePostDto;
-import com.yongfill.server.domain.posts.dto.DeletePostDto;
-import com.yongfill.server.domain.posts.dto.ReadPostDto;
-import com.yongfill.server.domain.posts.dto.UpdatePostDto;
+import com.yongfill.server.domain.posts.dto.post.CreatePostDto;
+import com.yongfill.server.domain.posts.dto.post.DeletePostDto;
+import com.yongfill.server.domain.posts.dto.post.ReadPostDto;
+import com.yongfill.server.domain.posts.dto.post.UpdatePostDto;
 import com.yongfill.server.domain.posts.entity.Category;
 import com.yongfill.server.domain.posts.entity.Post;
 import com.yongfill.server.domain.posts.repository.PostJpaRepository;
